Add docs, add HTTP recordings and other requested changes
