+++ docs/jupiterone.md
@@ -2,21 +2,9 @@
 
 ## Setup
 
-In this section, please provide details about how to set up the integration with
-JupiterOne. This may require provisioning some resources on the provider's side
-(perhaps a role, app, or api key) and passing information over to JupiterOne.
-
-## Data Model
-
-Provide an overview here of the resources collected from the integration. Please
-provide a mapping of how the resources collected map to the JupiterOne Data
-Model. The tables below were taken from the Azure integration to provide an
-example of how to display that information.
-
-When you start developing an integration, please clear out the tables below. As
-you add support for new entities and relationships, please update the tables and
-document the addition in the [CHANGELOG.md](../CHANGELOG.md) file at the root of
-the project.
+JupiterOne provides a managed integration for BambooHR. The integration connects
+directly to BambooHR API to obtain configuration metadata and analyze resource
+relationships.
 
 <!-- {J1_DOCUMENTATION_MARKER_START} -->
 <!--
@@ -40,7 +28,6 @@ The following entities are created:
 | Account   | `bamboohr_account` | `Account`       |
 | User      | `bamboohr_user`    | `User`          |
 | File      | `bamboohr_file`    | `DataObject`    |
-| Company   | `bamboohr_company` | `Organization`  |
 
 ### Relationships
 
@@ -50,8 +37,7 @@ The following relationships are created/mapped:
 | --------------------- | --------------------- | --------------------- |
 | `bamboohr_account`    | **HAS**               | `bamboohr_user`       |
 | `bamboohr_user`       | **HAS**               | `bamboohr_file`       |
-| `bamboohr_account`    | **HAS**               | `bamboohr_company`    |
-| `bamboohr_company`    | **HAS**               | `bamboohr_file`       |
+| `bamboohr_account`    | **HAS**               | `bamboohr_file`       |
 
 <!--
 ********************************************************************************
