+++ src/steps/index.ts
@@ -1,13 +1,11 @@
 import { accountSteps } from './account';
 import { accessSteps } from './access';
-import { companySteps } from './company';
 import { employeeFilesSteps } from './employee-files';
 import { companyFilesSteps } from './company-files';
 
 const integrationSteps = [
   ...accountSteps,
   ...accessSteps,
-  ...companySteps,
   ...employeeFilesSteps,
   ...companyFilesSteps,
 ];
